Add more data to json entities

+++ app/src/main/kotlin/domain/Progress.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Progress(
     val completion: Double?,
-    val printTime: Int?,
-    val printTimeLeft: Int?,
+    val printTime: Long?,
+    val printTimeLeft: Long?,
 )
